Assignment 1.3
I didn't need to change too much for this assignment which I think is pretty cool! :3 slowly getting there~

I don't like the image's settings so I'll definitely tinker with it, but the nav settings for the buttons to appear on the left of the page is just beautiful handiwork I'm glad I learned.